% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estmation.R
\name{Estmation}
\alias{Estmation}
\title{Estimating the error variance sigma^2 and the L_infinity bound M^2}
\usage{
Estmation(D, ntimes)
}
\arguments{
\item{D}{an N*(d+1) data frame. Full sample, the first d column
represents covariates and the last column represents response.}

\item{ntimes}{a positive integer. Number of repetitions.}
}
\value{
a ntimes*2 matrix. Each row represents an estimate of sigma^2 and M^2.
}
\description{
Estimating the error variance sigma^2 and the L_infinity bound M^2
}
\examples{
N = 1000
d = 2
n = 20
M2 = 10
sigma2 = 10
const = matrix(rep(1,N),ncol=1)
X = matrix(rnorm(N*d),ncol=d)
e = matrix(rnorm(N,sd=sqrt(sigma2)),ncol=1)
Y = const*2+X\%*\%matrix(c(3,3),ncol=1)+e
D = data.frame(X=X,Y=Y)
Estmation(D,10)
}
