% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/POPT.R
\name{POPT}
\alias{POPT}
\title{Calculating the optimal weight p*(r,d) of the minimax design}
\usage{
POPT(r, d)
}
\arguments{
\item{r}{a real number belongs to [0,1]. Relative importance of bias to variance.}

\item{d}{a positive integer. Dimension.}
}
\value{
the optimal weight p*(r,d) of the minimax design.
}
\description{
Calculating the optimal weight p*(r,d) of the minimax design
}
\examples{
POPT(0.5,1)
POPT(0.8,10)
}
