% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PPlot.R
\name{PPlot}
\alias{PPlot}
\title{Pair Plot of a design D}
\usage{
PPlot(D)
}
\arguments{
\item{D}{a n*d design matrix}
}
\value{
a pair plot.
}
\description{
Pair Plot of a design D
}
\examples{
D = p_MLHD(30,3,0.8)
PPlot(D)
}
