% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rpinv.R
\name{Rpinv}
\alias{Rpinv}
\title{Minimax transformation from [0,1]^d to [0,1]^d.}
\usage{
Rpinv(p, U)
}
\arguments{
\item{p}{a real number belongs to [0,1].}

\item{U}{a real matrix belongs to [0,1]^(n*d).}
}
\value{
a n*d real matrix that represents U after minimax transformation.
}
\description{
Minimax transformation from [0,1]^d to [0,1]^d.
}
\examples{
## An example of 30-point 2-dimensional random minimax design with p=0.5
n = 30
d = 2
p = 0.5
U = matrix(runif(n*d),ncol=d)
D = Rpinv(p,U)
PPlot(D)
}
